Use separate exception handler for CP
